Extracted assure_file_path_integrity function
